% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{copy_unify_outputs}
\alias{copy_unify_outputs}
\title{Copy Unify Output}
\usage{
copy_unify_outputs(
  rc3_directory = NULL,
  short_id = "test",
  unify_directory = NULL,
  references_directory = NULL
)
}
\arguments{
\item{rc3_directory}{where you want things copied to (sort of)}

\item{short_id}{the short id of the project}

\item{unify_directory}{where the unify outputs are stored}

\item{references_directory}{where the reference data is stored}
}
\value{
NULL, invisibly
}
\description{
Copy the output generated by the unify script, to the setup directory.
}
