% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rc3_check_gzip}
\alias{rc3_check_gzip}
\title{check gzip files}
\usage{
rc3_check_gzip(fasta, regexp = "fq.gz")
}
\arguments{
\item{fasta}{the directory to check, or a list of files directly.}
}
\description{
Runs \code{gzip -t} on a set of files to verify there is nothing wrong with
them before running pump and unify.
\code{gzip} must be installed and available on the system.
Because this is potentially making \strong{a lot} of \code{system2} calls via \code{purrr},
it is not easy to interrupt once begun.
}
