% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{rc3_copy_unify_outputs}
\alias{rc3_copy_unify_outputs}
\title{Copy Unify Output}
\usage{
rc3_copy_unify_outputs(
  rc3_directory = NULL,
  short_id = "test",
  unify_directory = NULL,
  references_directory = NULL,
  organism = "human",
  annotation = "G026"
)
}
\arguments{
\item{rc3_directory}{where you want things copied to (sort of)}

\item{short_id}{the short id of the project}

\item{unify_directory}{where the unify outputs are stored}

\item{references_directory}{where the reference data is stored}

\item{organism}{what organism is the data (for real or pretend)}

\item{annotation}{what annotations are we interested in (for real or pretend)}
}
\value{
output of the "tree" command on rc3_directory
}
\description{
Copy the output generated by the unify script, to the setup directory,
and symlink a few things. Note that \code{recount3} is expecting \emph{human} or \emph{mouse},
with the relevant annotation. If you are using this, you could have generated
anything you wanted (I build this to handle \emph{zebrafish} using \emph{MZ11}). Therefore,
it copies the files, then symlinks them to the names that \code{recount3} expects.
}
