% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pump_unify.R
\name{rc3_run_pump_unify}
\alias{rc3_run_pump_unify}
\title{run recount-pump and recount-unify}
\usage{
rc3_run_pump_unify(
  fasta = getwd(),
  outputs = fs::path_dir(fs::path_expand(fasta)),
  monorail = "monorail-external",
  recount_pump = "recount-pump_1.1.3.sif",
  recount_unify = "recount-unify_1.1.1.sif",
  reference_path = ".",
  reference = "hg38",
  studyid = "other1",
  shortid = "test",
  ncore = 1,
  run = "both",
  delete_oldpump = "yes",
  run_or_show = "run"
)
}
\arguments{
\item{fasta}{the directory with the input samples to be run}

\item{outputs}{where to keep outputs, assumed to be the parent of fasta}

\item{monorail}{where is the monorail repository}

\item{recount_pump}{where is the recount-pump sif image}

\item{recount_unify}{where is the recount-unify sif image}

\item{studyid}{the study id}

\item{shortid}{short id}

\item{ncore}{how many cores to use}

\item{run}{which parts of the workflow to run? (both, pump, unify)}

\item{delete_oldpump}{should the old recount-pump results be deleted and regenerated?}

\item{run_or_show}{do you want the commands shown, or actually run?}
}
\value{
unify output directory
}
\description{
Given the parameters, attempts to create and run the commands to carry out the
runs of recount-pump and recount-unify on a set of samples. Assumes that all of
the samples in the directory of \code{fasta} are all associated with the same \code{study_id}
in the \code{studyid} argument, and creates the \code{sample_metadata.tsv} file for the \code{unify} step.
}
\details{
An important thing to know, if your samples end with \verb{_1} or \verb{_4} etc, you want to rename
them to have two numbers at the end, like \verb{_01} or \verb{_04}. The \code{recount3} workflow does not
like underscores at the end of a sample identifier. \verb{_1.fq.gz} and \verb{_2.fq.gz} for paired-end
reads is normal and expected, it's the bits before that that are important.
}
