% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{rc3_setup_directory}
\alias{rc3_setup_directory}
\title{setup a new directory}
\usage{
rc3_setup_directory(base_dir = ".", short_id = "test", organism = "human")
}
\arguments{
\item{base_dir}{the top level directory where projects should live}

\item{short_id}{the short identifier used when unifying your outputs}
}
\value{
project directory list
}
\description{
Sets up directory structure needed for recount3 to read data in.
Note that \code{recount3} only currently works with \emph{human} or \emph{mouse},
so regardless your \emph{actual} organism you used, you must choose
one of those two to make this work.
}
