% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_rc3_directory}
\alias{setup_rc3_directory}
\title{setup a new directory}
\usage{
setup_rc3_directory(base_dir = ".", short_id = "test")
}
\arguments{
\item{base_dir}{the top level directory where projects should live}

\item{short_id}{the short identifier used when unifying your outputs}
}
\value{
project directory list
}
\description{
Sets up directory structure needed for recount3 to read data in.
}
