% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pump_unify.R
\name{rc3_run_pump_unify}
\alias{rc3_run_pump_unify}
\title{run recount-pump and recount-unify}
\usage{
rc3_run_pump_unify(
  fasta = getwd(),
  outputs = fs::path_dir(fs::path_expand(fasta)),
  monorail = "monorail-external",
  recount_pump = "recount-pump_1.1.3.sif",
  recount_unify = "recount-pump_1.1.3.sif",
  reference_path = ".",
  reference = "hg38",
  studyid = "other1",
  shortid = "test",
  ncore = 1
)
}
\arguments{
\item{fasta}{the directory with the input samples to be run}

\item{outputs}{where to keep outputs, assumed to be the parent of fasta}

\item{monorail}{where is the monorail repository}

\item{recount_pump}{where is the recount-pump sif image}

\item{recount_unify}{where is the recount-unify sif image}

\item{studyid}{the study id}

\item{shortid}{short id}

\item{ncore}{how many cores to use}
}
\value{
unify output directory
}
\description{
Given the parameters, attempts to create and run the commands to carry out the
runs of recount-pump and recount-unify on a set of samples.
}
